+++ core/src/bms/player/beatoraja/MainController.java
@@ -42,6 +42,8 @@ import bms.player.beatoraja.song.*;
 import bms.player.beatoraja.stream.StreamController;
 import bms.tool.mdprocessor.MusicDownloadProcessor;
 
+import static bms.player.beatoraja.skin.SkinProperty.*;
+
 /**
  * アプリケーションのルートクラス
  *
@@ -128,6 +130,8 @@ public class MainController extends ApplicationAdapter {
 	private final Array<MainStateListener> stateListener = new Array<MainStateListener>();
 
 	public ImGuiRenderer imGui;
+
+	public List<IRSendStatus> irSendStatus = new ArrayList<IRSendStatus>();
 
 	public MainController(Path f, Config config, PlayerConfig player, BMSPlayerMode auto, boolean songUpdated) {
 		this.auto = auto;
@@ -422,6 +426,34 @@ public class MainController extends ApplicationAdapter {
 
 		if(ir.length > 0) {
 			messageRenderer.addMessage(ir.length + " IR Connection Succeed" ,5000, Color.GREEN, 1);
+
+			Thread irResendProcess = new Thread(() -> {
+				for (;;) {
+					final long now = System.currentTimeMillis();
+						try {
+							List<IRSendStatus> removeIrSendStatus = new ArrayList<IRSendStatus>();
+
+							for(IRSendStatus irc : irSendStatus) {
+								long timeUntilNextTry = (long)(Math.pow(4, irc.retry) * 1000);
+								if (irc.retry != 0 && now - irc.lastTry >= timeUntilNextTry) {
+									irc.send();
+								}
+								if(irc.retry < 0 || irc.retry > getConfig().getIrSendCount()) {
+									removeIrSendStatus.add(irc);
+								}
+							}
+							irSendStatus.removeAll(removeIrSendStatus);
+
+							try {
+								Thread.sleep(3000, 0);
+							} catch (InterruptedException e) {
+							}
+						} catch (Exception e) {
+							Logger.getGlobal().severe(e.getMessage());
+						}
+				}
+			});
+			irResendProcess.start();
 		}
 	}
 
@@ -946,6 +978,36 @@ public class MainController extends ApplicationAdapter {
 			this.config = config;
 			this.connection = connection;
 			this.player = player;
+		}
+	}
+
+	public static class IRSendStatus {
+		public final IRConnection ir;
+		public final SongData song;
+		public final ScoreData score;
+		public int retry = 0;
+		public long lastTry = 0;
+
+		public IRSendStatus(IRConnection ir, SongData song, ScoreData score) {
+			this.ir = ir;
+			this.song = song;
+			this.score = score;
+		}
+
+		public boolean send() {
+			Logger.getGlobal().info("IRへスコア送信中 : " + song.getTitle());
+			lastTry = System.currentTimeMillis();
+			IRResponse<Object> send1 = ir.sendPlayData(new IRChartData(song), new bms.player.beatoraja.ir.IRScoreData(score));
+			if(send1.isSucceeded()) {
+				Logger.getGlobal().info("IRスコア送信完了 : " + song.getTitle());
+				retry = -255;
+				return true;
+			} else {
+				Logger.getGlobal().warning("IRスコア送信失敗 : " + send1.getMessage());
+				retry++;
+				return false;
+			}
+
 		}
 	}
 }
